Refactoring state on provider
