+++ src/pages/Provider/Provider.layout.vue
@@ -3,11 +3,12 @@
     <provider-header
       :logo-url="provider.logo_url"
       :name="provider.commercial_name"
+      @onSearch="$emit('onSearch')"
     />
     <div class="Provider__container">
       <suspense>
         <template #default>
-          <provider-products :id="id" :products="provider.products" />
+          <provider-products :products="products" />
         </template>
         <template #fallback>
           <base-loading />
@@ -18,13 +19,12 @@
 </template>
 
 <script lang="ts">
-import { defineAsyncComponent, defineComponent } from 'vue';
+import { defineComponent, ref } from 'vue';
 import { BaseLoading } from '../../components/LoadingComponent';
+import { getProductsByProvider } from '../../actions';
 import providerHeader from './ProviderHeader.vue';
-
-const ProviderProducts = defineAsyncComponent(
-  () => import('./ProviderProducts.vue')
-);
+import ProviderProducts from './ProviderProducts.vue';
+import { product } from 'src/utils';
 
 export default defineComponent({
   name: 'ProviderLayout',
@@ -40,11 +40,26 @@ export default defineComponent({
       },
     },
   },
-  emits: ['update:tabValue'],
+  emits: ['onSearch'],
   components: {
     ProviderProducts,
     BaseLoading,
     providerHeader,
+  },
+  async setup(props) {
+    const products = ref<product[]>([]);
+
+    if (props.id && typeof props.id === 'string') {
+      await getProductsByProvider(props.id).then((res) => {
+        console.log(res);
+        products.value = res;
+      });
+    }
+
+    return {
+      getProductsByProvider,
+      products,
+    };
   },
 });
 </script>
