+++ src/pages/Provider/ProviderProducts.vue
@@ -9,34 +9,22 @@
 </template>
 
 <script lang="ts">
-import { defineComponent, ref } from 'vue';
-import { getProductsByProvider } from '../../actions';
+import { product } from 'src/utils';
+import { defineComponent, PropType } from 'vue';
 import ProviderProduct from './ProviderProduct.vue';
 
 export default defineComponent({
   name: 'ProviderProducts',
   props: {
-    id: {
-      type: String,
-      default: '',
+    products: {
+      type: Array as PropType<product[]>,
+      default: () => {
+        return [];
+      },
     },
   },
   components: {
     ProviderProduct,
-  },
-  async setup(props) {
-    const products = ref({});
-
-    if (props.id && typeof props.id === 'string') {
-      await getProductsByProvider(props.id).then((res) => {
-        console.log(res);
-        products.value = res;
-      });
-    }
-
-    return {
-      products,
-    };
   },
 });
 </script>
